# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TaxWise AI - W-2 Tax Return Automation API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai-agent/config:
    get:
      operationId: getAiAgentConfig
      tags: [aiAgent]
      summary: Get current AI agent configuration
      responses:
        "200":
          description: AI agent configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiAgentConfig"
    put:
      operationId: updateAiAgentConfig
      tags: [aiAgent]
      summary: Update AI agent configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAiAgentConfigBody"
      responses:
        "200":
          description: Updated config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiAgentConfig"
  /ai-agent/datasets:
    get:
      operationId: listTrainingDatasets
      tags: [aiAgent]
      summary: List training datasets
      responses:
        "200":
          description: List of datasets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrainingDataset"
    post:
      operationId: createTrainingDataset
      tags: [aiAgent]
      summary: Create a new training dataset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTrainingDatasetBody"
      responses:
        "201":
          description: Created dataset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingDataset"
  /ai-agent/datasets/{id}/scrape:
    post:
      operationId: scrapeDataForDataset
      tags: [aiAgent]
      summary: Trigger web scraping to add training data to a dataset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScrapeDataBody"
      responses:
        "200":
          description: Scraping job started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScrapeJobResult"
  /ai-agent/train:
    post:
      operationId: startTrainingJob
      tags: [aiAgent]
      summary: Start a training job with a dataset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartTrainingJobBody"
      responses:
        "202":
          description: Training job started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingJob"
...
  /knowledge-graph/entities:
    get:
      operationId: listEntities
      tags: [knowledgeGraph]
      summary: List entities in the knowledge graph
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: List of entities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GraphEntity"
    post:
      operationId: createEntity
      tags: [knowledgeGraph]
      summary: Create a new entity in the knowledge graph
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEntityBody"
      responses:
        "201":
          description: Created entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphEntity"
  /knowledge-graph/entities/{id}:
    get:
      operationId: getEntity
      tags: [knowledgeGraph]
      summary: Get an entity with its connections
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Entity with connections
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphEntityWithConnections"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    put:
      operationId: updateEntity
      tags: [knowledgeGraph]
      summary: Update an entity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEntityBody"
      responses:
        "200":
          description: Updated entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphEntity"
    delete:
      operationId: deleteEntity
      tags: [knowledgeGraph]
      summary: Delete an entity and its connections
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /knowledge-graph/connections:
    get:
      operationId: listConnections
      tags: [knowledgeGraph]
      summary: List connections between entities
      responses:
        "200":
          description: List of connections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GraphConnection"
    post:
      operationId: createConnection
      tags: [knowledgeGraph]
      summary: Create a connection between two entities
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConnectionBody"
      responses:
        "201":
          description: Created connection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphConnection"
  /knowledge-graph/graph:
    get:
      operationId: getFullGraph
      tags: [knowledgeGraph]
      summary: Get full graph data for visualization
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: entityType
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Full graph data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphData"
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    AiAgentConfig:
      type: object
      required: [primaryModel, temperature, maxTokens, enableOcr, enableWebScraping]
      properties:
        primaryModel:
          type: string
        fallbackModel:
          type: [string, 'null']
        temperature:
          type: number
        maxTokens:
          type: integer
        enableOcr:
          type: boolean
        enableWebScraping:
          type: boolean
        systemPrompt:
          type: [string, 'null']
    UpdateAiAgentConfigBody:
      type: object
      properties:
        primaryModel:
          type: string
        fallbackModel:
          type: string
        temperature:
          type: number
        maxTokens:
          type: integer
        enableOcr:
          type: boolean
        enableWebScraping:
          type: boolean
        systemPrompt:
          type: string
    TrainingDataset:
      type: object
      required: [id, name, type, status, recordCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: [string, 'null']
        type:
          type: string
          enum: [structured, unstructured, mixed, scraped]
        status:
          type: string
          enum: [active, processing, archived]
        recordCount:
          type: integer
        objectPath:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
    CreateTrainingDatasetBody:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [structured, unstructured, mixed, scraped]
        objectPath:
          type: string
    ScrapeDataBody:
      type: object
      required: [urls]
      properties:
        urls:
          type: array
          items:
            type: string
            format: uri
        maxDepth:
          type: integer
          default: 1
        selectors:
          type: array
          items:
            type: string
    ScrapeJobResult:
      type: object
      required: [jobId, status, urlsQueued]
      properties:
        jobId:
          type: string
        status:
          type: string
          enum: [queued, running, completed, failed]
        urlsQueued:
          type: integer
    StartTrainingJobBody:
      type: object
      required: [datasetId]
      properties:
        datasetId:
          type: integer
        modelName:
          type: string
        hyperparameters:
          type: object
          additionalProperties: true
    TrainingJob:
      type: object
      required: [id, datasetId, status, createdAt]
      properties:
        id:
          type: integer
        datasetId:
          type: integer
        modelName:
          type: [string, 'null']
        status:
          type: string
          enum: [queued, running, completed, failed]
        progress:
          type: [number, 'null']
        metrics:
          type: object
          additionalProperties: true
        errorMessage:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: [string, 'null']
          format: date-time
    GraphEntity:
      type: object
      required: [id, name, type, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        description:
          type: [string, 'null']
        properties:
          type: object
          additionalProperties: true
        color:
          type: [string, 'null']
        size:
          type: [number, 'null']
        connectionCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    GraphEntityWithConnections:
      allOf:
        - $ref: "#/components/schemas/GraphEntity"
        - type: object
          properties:
            connections:
              type: array
              items:
                $ref: "#/components/schemas/GraphConnection"
    CreateEntityBody:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
        description:
          type: string
        properties:
          type: object
          additionalProperties: true
        color:
          type: string
        size:
          type: number
    UpdateEntityBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        description:
          type: string
        properties:
          type: object
          additionalProperties: true
        color:
          type: string
        size:
          type: number
    GraphConnection:
      type: object
      required: [id, sourceId, targetId, relationshipType, createdAt]
      properties:
        id:
          type: integer
        sourceId:
          type: integer
        targetId:
          type: integer
        relationshipType:
          type: string
        weight:
          type: [number, 'null']
        properties:
          type: object
          additionalProperties: true
        sourceName:
          type: [string, 'null']
        targetName:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
    CreateConnectionBody:
      type: object
      required: [sourceId, targetId, relationshipType]
      properties:
        sourceId:
          type: integer
        targetId:
          type: integer
        relationshipType:
          type: string
        weight:
          type: number
        properties:
          type: object
          additionalProperties: true
    GraphData:
      type: object
      required: [nodes, edges]
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/GraphEntity"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/GraphConnection"
        stats:
          type: object
          properties:
            totalNodes:
              type: integer
            totalEdges:
              type: integer
            clusters:
              type: integer
...